refactor: move ad-hoc test scripts to tests/scripts/
Relocate 14 ad-hoc test scripts out of the pytest discovery path.
These scripts require live credentials or block indefinitely and are
unsafe to run under `pytest`. They remain available for manual use.
