+++ tests/scripts/test_auto_provisioning.py
