+++ tests/scripts/test_streaming.py
