+++ tests/scripts/test_trading.py
